fix in `html` - register the right event handlers with the right elems:
Create a `html.funcs` as well as a `html.i`, and use those
as references to the functions to be inserted.

This fixes Component so that the new stuff in
examples/recipes.js works OK.

NOTE: `html` will detect a statful Component, and will only
retrieve its view, **not** run its rull setState()/render()
lifecycle. This has a number of implications:

- better performance
- parent component holds the only state
- parent component inly triggers re-renders
- nested stateful components have a undefined `.container`
  - therefore calling setState() of child components does nothing